+++ src/src/component/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'unstated';
-import Header from './component/screen/common/Header';
-import Main from './component/screen/common/Main';
-import Footer from './component/screen/common/Footer';
+import Header from '../common/Header';
+import Main from '../common/Main';
+import Footer from '../common/Footer';
 
 function App() {
   return (
