+++ src/src/component/Profile.js
@@ -6,11 +6,11 @@ import {
   MuiThemeProvider,
   makeStyles,
 } from '@material-ui/core';
-import Tooltip from '../../common/Tooltip';
+import Tooltip from '../common/Tooltip';
 import { Subscribe } from 'unstated';
-import AppContainer from '../../../container/AppContainer';
-import avatar from '../../../../image/sample/avatar.png';
-import theme from '../../../asset/Theme';
+import AppContainer from '../container/AppContainer';
+import avatar from '../asset/image/avatar.png';
+import theme from '../asset/Theme';
 
 const useStyles = makeStyles({
   avatar: {
